requirements.txt
langchain_community 
python-dotenv
langchain_google_genai
langchain_core